Refactors file names
