+++ multibody/hydroelastics/test/contact_surface_from_level_set_test.cc
@@ -1,4 +1,4 @@
-#include "drake/multibody/hydroelastics/contact_surface_calculator.h"
+#include "drake/multibody/hydroelastics/contact_surface_from_level_set.h"
 
 #include <fstream>
